+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 1.4.1 (February 9, 2023)
+
 ### Adds
 
 - Adds the `pattern` prop to the `TextInput` component.